feat [M1953] DRY types
