+++ src/types/protocols.ts
@@ -1,0 +1,8 @@
+export type Protocol =
+  | 'fishbelt'
+  | 'benthicpit'
+  | 'benthiclit'
+  | 'benthicpqt'
+  | 'bleachingqc'
+  | 'habitatcomplexity'
+  | 'macroinvertebrate'
